requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-WTF
WTForms
email-validator
PyMySQL
python-dotenv